imageAnalysis/imageCaptioning_phi4_delete/requirements.txt
accelerate
#diffusers
#invisible_watermark
torch
transformers
#xformers
gradio>=4.0.0
Pillow